requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==2.5.2
httpx==0.25.2
jinja2==3.1.2
python-jose==3.3.0
cryptography==41.0.5
beautifulsoup4==4.12.2
reportlab         # PDF üretimi
schedule          # Basit cron alternatifi (isteğe bağlı)
